+++ extensions/blueidea/browser/index.js
@@ -126,62 +126,7 @@ const resetObj = {
 };
 
 // This element contains the styling for the module
-var styleObject = {
-  ventil_forbundet: {
-    radius: 8,
-    fillColor: "#00ff00",
-    color: "#000",
-    weight: 1,
-    opacity: 1,
-    fillOpacity: 0.8,
-  },
-  ventil: {
-    radius: 5,
-    fillColor: "#ff7800",
-    color: "#000",
-    weight: 1,
-    opacity: 1,
-    fillOpacity: 0.8,
-  },
-  selectedLedning: {
-    color: "#AA4A44",
-    weight: 8,
-  },
-  selectedPoint: {
-    html: `
-  <svg width="24" height="24" xmlns="http://www.w3.org/2000/svg" fill-rule="evenodd" clip-rule="evenodd">
-  <path d="M12 11.293l10.293-10.293.707.707-10.293 10.293 10.293 10.293-.707.707-10.293-10.293-10.293 10.293-.707-.707 10.293-10.293-10.293-10.293.707-.707 10.293 10.293z"/>
-  </svg>
-  `,
-    className: "",
-    iconSize: [24, 24], // size of the icon
-    //iconAnchor: [-10, -10], // point of the icon which will correspond to marker's location
-  },
-  matrikel: {
-    color: "#000000",
-    weight: 1,
-    opacity: 1,
-    fillOpacity: 0.2,
-    dashArray: "5,3",
-  },
-  buffer: {
-    color: "#ff7800",
-    weight: 1,
-    opacity: 1,
-    fillOpacity: 0.1,
-    dashArray: "5,3",
-  },
-  alarmPosition: {
-    html: `
-  <svg fill="#000000" width="24px" height="24px" viewBox="0 0 57.6 57.6" xmlns="http://www.w3.org/2000/svg">
-  <path d="M28.709 0c-10.872 0 -19.71 8.838 -19.71 19.706 0 5.418 2.408 10.436 7.362 15.347 7.193 7.139 10.548 13.73 10.548 20.747v1.8h3.6v-1.8c0 -6.984 3.308 -13.385 10.728 -20.743 4.954 -4.914 7.362 -9.932 7.362 -15.35 0 -10.868 -8.838 -19.706 -19.89 -19.706" fill-rule="evenodd"/>
-  </svg>
-  `,
-    className: "",
-    iconSize: [24, 24], // size of the icon
-    iconAnchor: [12, 24], // point of the icon which will correspond to marker's location
-  },
-};
+require("./style.js");
 
 /**
  * async function to query matrikel inside a single buffer
@@ -301,128 +246,7 @@ module.exports = {
      *
      * @type {*|exports|module.exports}
      */
-    var dict = {
-      Info: {
-        da_DK:
-          "BlueIdea / Lukkeliste er et modul der hjælper med at finde relevante adresser til en lukning. Du kan tegne et område på kortet, og få en liste over adresser i området. Du kan også udpege et punkt på ledningsnettet, og få en liste af adresser der er forbundet til det punkt. Listerne med adresser kan derefter bruges til at sende beskeder ud med BlueIdea (Kræver aftale med BlueIdea).",
-        en_US:
-          "BlueIdea / Valve list is a module that helps to find relevant addressess for a valve closure. You can draw an area on the map, and get a list of addresses in the area. You can also select a point on the network, and get a list of addresses connected to that point. The lists of addresses can then be used to send messages with BlueIdea (Requires agreement with BlueIdea).",
-      },
-      "Plugin Tooltip": {
-        da_DK: "BlueIdea / Lukkeliste",
-        en_US: "BlueIdea / Valve list",
-      },
-      MissingLogin: {
-        da_DK:
-          "NB: Du skal være logget ind for at kunne bruge funktionen, og din konfiguration skal indeholde et brugerid.",
-        en_US:
-          "Please log in and set a user id in the configuration file to use this function",
-      },
-      Login: {
-        da_DK: "Log ind",
-        en_US: "Log in",
-      },
-      "Select point on map": {
-        da_DK: "Udpeg punkt på ledningsnet",
-        en_US: "Select point on map",
-      },
-      "Select point for alarmkabel": {
-        da_DK: "Beregn alarm",
-        en_US: "Calculate alarm",
-      },
-      "Found parcels": {
-        da_DK: "Fundne matrikler i område",
-        en_US: "Found parcels in area",
-      },
-      "Found addresses": {
-        da_DK: "Fundet adresser i matrikler",
-        en_US: "Found addresses in parcels",
-      },
-      "Error in search": {
-        da_DK: "Fejl i søgning",
-        en_US: "Error in search",
-      },
-      "Draw area": {
-        da_DK: "Tegn områder",
-        en_US: "Draw area",
-      },
-      "Select area": {
-        da_DK: "Udpeg",
-        en_US: "Select",
-      },
-      "Show results": {
-        da_DK: "Resultater",
-        en_US: "Results",
-      },
-      "Waiting to start": {
-        da_DK: "Henter jordstykker",
-        en_US: "Fetching parcels",
-      },
-      "Go to blueidea": {
-        da_DK: "Opret, og gå til blueidea",
-        en_US: "Create and go to blueidea",
-      },
-      "Valve list": {
-        da_DK: "Lukkeliste",
-        en_US: "Valve list",
-      },
-      "Too many features selected": {
-        da_DK: "For mange steder at lede efter adresser!",
-        en_US: "Too many places to look for addresses!",
-      },
-      "Download addresses": {
-        da_DK: "Download adresser",
-        en_US: "Download addresses",
-      },
-      "Get addresses": {
-        da_DK: "Hent adresser",
-        en_US: "Get addresses",
-      },
-      "Download valves": {
-        da_DK: "Download ventil-liste",
-        en_US: "Download valve list",
-      },
-      NotAllowedBlueIdea: {
-        da_DK: "Du har ikke adgang til BlueIdea",
-        en_US: "You are not allowed to use BlueIdea",
-      },
-      "Starting analysis": {
-        da_DK: "Analyserer",
-        en_US: "Analyzing",
-      },
-      "modify parcels": {
-        da_DK: "Tilføj eller fjern matrikler",
-        en_US: "Add or remove parcels",
-      },
-      "Select profile": {
-        da_DK: "Vælg profil",
-        en_US: "Select profile",
-      },
-      "Alarm cable": {
-        da_DK: "Alarmkabel",
-        en_US: "Alarm cable",
-      },
-      "Distance": {
-        da_DK: "Afstand",
-        en_US: "Distance",
-      },
-      "Distance not set": {
-        da_DK: "Ugyldig afstand",
-        en_US: "Invalid distance",
-      },
-      "Alarm found": {
-        da_DK: "Mulige placeringer fundet",
-        en_US: "Possible alarms found",
-      },
-      "Lukkeliste is ready": {
-        da_DK: "Lukkeliste klar",
-        en_US: "Valve list ready",
-      },
-      "Lukkeliste not ready": {
-        da_DK: "Lukkeliste ikke klar",
-        en_US: "Valve list not ready",
-      },
-    };
+    require("./i8n.js");
 
     /**
      *
@@ -1556,7 +1380,10 @@ module.exports = {
         for (let i = 0; i < matrikler.features.length; i++) {
           let feature = matrikler.features[i];
           results.push(await findAddressesInMatrikel(feature));
-          me.createSnack(__("Found addresses") + " " + i + "/" + matrikler.features.length);
+          // Show progress per 25 features
+          if (i % 25 == 0) {
+            me.createSnack(__("Found addresses") + " " + i + "/" + matrikler.features.length);
+          }
         }
 
         let adresser = this.mergeAdresser(results);
